aggregates: rename cross_metric to aggregates
This change reorganizes aggregates code.

* gnocchi.aggregates becomes gnocchi.deprecated_aggregates
* gnocchi.rest.cross_metric becomes gnocchi.rest.aggregates.processor

Related #419
