+++ gnocchi/deprecated_aggregates/__init__.py
