+++ gnocchi/tests/test_aggregates.py
@@ -1,6 +1,6 @@
 # -*- encoding: utf-8 -*-
 #
-# Copyright 2014-2015 OpenStack Foundation
+# Copyright © 2014-2016 eNovance
 #
 # Licensed under the Apache License, Version 2.0 (the "License"); you may
 # not use this file except in compliance with the License. You may obtain
@@ -14,96 +14,948 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 import datetime
+import functools
 import uuid
 
 import numpy
-from stevedore import extension
 
-from gnocchi import aggregates
-from gnocchi.aggregates import moving_stats
+from gnocchi import carbonara
 from gnocchi import incoming
 from gnocchi import indexer
-from gnocchi.tests import base as tests_base
-from gnocchi.tests import utils as tests_utils
-from gnocchi import utils
+from gnocchi.rest.aggregates import processor
+from gnocchi import storage
+from gnocchi.tests import base
 
 
-class TestAggregates(tests_base.TestCase):
+def datetime64(*args):
+    return numpy.datetime64(datetime.datetime(*args))
 
+
+class TestAggregatedTimeseries(base.BaseTestCase):
+    @staticmethod
+    def _resample_and_merge(ts, agg_dict):
+        """Helper method that mimics _add_measures workflow."""
+        grouped = ts.group_serie(agg_dict['sampling'])
+        existing = agg_dict.get('return')
+        agg_dict['return'] = carbonara.AggregatedTimeSerie.from_grouped_serie(
+            grouped, agg_dict['sampling'], agg_dict['agg'],
+            max_size=agg_dict.get('size'), truncate=True)
+        if existing:
+            existing.merge(agg_dict['return'])
+            agg_dict['return'] = existing
+
+    def test_aggregated_different_archive_no_overlap(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 50, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 50, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        tsb1.set_values(numpy.array([(datetime64(2014, 1, 1, 11, 46, 4), 4)],
+                                    dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+                        before_truncate_callback=functools.partial(
+                            self._resample_and_merge, agg_dict=tsc1))
+        tsb2.set_values(numpy.array([(datetime64(2014, 1, 1, 9, 1, 4), 4)],
+                                    dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+                        before_truncate_callback=functools.partial(
+                            self._resample_and_merge, agg_dict=tsc2))
+
+        dtfrom = datetime64(2014, 1, 1, 11, 0, 0)
+        self.assertRaises(processor.UnAggregableTimeseries,
+                          processor.aggregated,
+                          [tsc1['return'], tsc2['return']],
+                          from_timestamp=dtfrom, aggregation='mean')
+
+    def test_aggregated_different_archive_no_overlap2(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 50, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = carbonara.AggregatedTimeSerie(
+            sampling=numpy.timedelta64(60, 's'),
+            max_size=50,
+            aggregation_method='mean')
+
+        tsb1.set_values(numpy.array([(datetime64(2014, 1, 1, 12, 3, 0), 4)],
+                                    dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+                        before_truncate_callback=functools.partial(
+                            self._resample_and_merge, agg_dict=tsc1))
+        self.assertRaises(processor.UnAggregableTimeseries,
+                          processor.aggregated,
+                          [tsc1['return'], tsc2], aggregation='mean')
+
+    def test_aggregated_different_archive_overlap(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        # NOTE(sileht): minute 8 is missing in both and
+        # minute 7 in tsc2 too, but it looks like we have
+        # enough point to do the aggregation
+        tsb1.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 11, 0, 0), 4),
+            (datetime64(2014, 1, 1, 12, 1, 0), 3),
+            (datetime64(2014, 1, 1, 12, 2, 0), 2),
+            (datetime64(2014, 1, 1, 12, 3, 0), 4),
+            (datetime64(2014, 1, 1, 12, 4, 0), 2),
+            (datetime64(2014, 1, 1, 12, 5, 0), 3),
+            (datetime64(2014, 1, 1, 12, 6, 0), 4),
+            (datetime64(2014, 1, 1, 12, 7, 0), 10),
+            (datetime64(2014, 1, 1, 12, 9, 0), 2)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc1))
+
+        tsb2.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 1, 0), 3),
+            (datetime64(2014, 1, 1, 12, 2, 0), 4),
+            (datetime64(2014, 1, 1, 12, 3, 0), 4),
+            (datetime64(2014, 1, 1, 12, 4, 0), 6),
+            (datetime64(2014, 1, 1, 12, 5, 0), 3),
+            (datetime64(2014, 1, 1, 12, 6, 0), 6),
+            (datetime64(2014, 1, 1, 12, 9, 0), 2),
+            (datetime64(2014, 1, 1, 12, 11, 0), 2),
+            (datetime64(2014, 1, 1, 12, 12, 0), 2)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc2))
+
+        dtfrom = datetime64(2014, 1, 1, 12, 0, 0)
+        dtto = datetime64(2014, 1, 1, 12, 10, 0)
+
+        # By default we require 100% of point that overlap
+        # so that fail
+        self.assertRaises(processor.UnAggregableTimeseries,
+                          processor.aggregated,
+                          [tsc1['return'], tsc2['return']],
+                          from_timestamp=dtfrom,
+                          to_timestamp=dtto, aggregation='mean')
+
+        # Retry with 80% and it works
+        output = processor.aggregated([
+            tsc1['return'], tsc2['return']],
+            from_timestamp=dtfrom, to_timestamp=dtto,
+            aggregation='mean', needed_percent_of_overlap=80.0)
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 12, 1, 0),
+             numpy.timedelta64(60, 's'), 3.0),
+            (datetime64(2014, 1, 1, 12, 2, 0),
+             numpy.timedelta64(60, 's'), 3.0),
+            (datetime64(2014, 1, 1, 12, 3, 0),
+             numpy.timedelta64(60, 's'), 4.0),
+            (datetime64(2014, 1, 1, 12, 4, 0),
+             numpy.timedelta64(60, 's'), 4.0),
+            (datetime64(2014, 1, 1, 12, 5, 0),
+             numpy.timedelta64(60, 's'), 3.0),
+            (datetime64(2014, 1, 1, 12, 6, 0),
+             numpy.timedelta64(60, 's'), 5.0),
+            (datetime64(2014, 1, 1, 12, 7, 0),
+             numpy.timedelta64(60, 's'), 10.0),
+            (datetime64(2014, 1, 1, 12, 9, 0),
+             numpy.timedelta64(60, 's'), 2.0),
+        ], list(output))
+
+    def test_aggregated_different_archive_overlap_edge_missing1(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        tsb1.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 3, 0), 9),
+            (datetime64(2014, 1, 1, 12, 4, 0), 1),
+            (datetime64(2014, 1, 1, 12, 5, 0), 2),
+            (datetime64(2014, 1, 1, 12, 6, 0), 7),
+            (datetime64(2014, 1, 1, 12, 7, 0), 5),
+            (datetime64(2014, 1, 1, 12, 8, 0), 3)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc1))
+
+        tsb2.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 11, 0, 0), 6),
+            (datetime64(2014, 1, 1, 12, 1, 0), 2),
+            (datetime64(2014, 1, 1, 12, 2, 0), 13),
+            (datetime64(2014, 1, 1, 12, 3, 0), 24),
+            (datetime64(2014, 1, 1, 12, 4, 0), 4),
+            (datetime64(2014, 1, 1, 12, 5, 0), 16),
+            (datetime64(2014, 1, 1, 12, 6, 0), 12)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc2))
+
+        # By default we require 100% of point that overlap
+        # but we allow that the last datapoint is missing
+        # of the precisest granularity
+        output = processor.aggregated([
+            tsc1['return'], tsc2['return']], aggregation='sum')
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 12, 3, 0),
+             numpy.timedelta64(60, 's'), 33.0),
+            (datetime64(2014, 1, 1, 12, 4, 0),
+             numpy.timedelta64(60, 's'), 5.0),
+            (datetime64(2014, 1, 1, 12, 5, 0),
+             numpy.timedelta64(60, 's'), 18.0),
+            (datetime64(2014, 1, 1, 12, 6, 0),
+             numpy.timedelta64(60, 's'), 19.0),
+        ], list(output))
+
+    def test_aggregated_different_archive_overlap_edge_missing2(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        tsb1.set_values(numpy.array([(datetime64(2014, 1, 1, 12, 3, 0), 4)],
+                                    dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+                        before_truncate_callback=functools.partial(
+                            self._resample_and_merge, agg_dict=tsc1))
+
+        tsb2.set_values(numpy.array([(datetime64(2014, 1, 1, 11, 0, 0), 4),
+                                     (datetime64(2014, 1, 1, 12, 3, 0), 4)],
+                                    dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+                        before_truncate_callback=functools.partial(
+                            self._resample_and_merge, agg_dict=tsc2))
+
+        output = processor.aggregated(
+            [tsc1['return'], tsc2['return']], aggregation='mean')
+        self.assertEqual([
+            (datetime64(
+                2014, 1, 1, 12, 3, 0
+            ), numpy.timedelta64(60000000000, 'ns'), 4.0),
+        ], list(output))
+
+    def test_processor_with_random_holes_derived_boundaries(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        tsb1.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 1, 0), 9),
+            (datetime64(2014, 1, 1, 12, 2, 0), 1),
+            (datetime64(2014, 1, 1, 12, 4, 0), 5),
+            (datetime64(2014, 1, 1, 12, 6, 0), 3)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc1))
+
+        tsb2.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 0, 0), 6),
+            (datetime64(2014, 1, 1, 12, 1, 0), 2),
+            (datetime64(2014, 1, 1, 12, 2, 0), 13),
+            (datetime64(2014, 1, 1, 12, 3, 0), 24),
+            (datetime64(2014, 1, 1, 12, 4, 0), 4)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc2))
+
+        output = processor.aggregated([
+            tsc1['return'], tsc2['return']], aggregation='mean',
+            needed_percent_of_overlap=50.0)
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 12, 1, 0),
+             numpy.timedelta64(60000000000, 'ns'), 5.5),
+            (datetime64(2014, 1, 1, 12, 2, 0),
+             numpy.timedelta64(60000000000, 'ns'), 7.0),
+            (datetime64(2014, 1, 1, 12, 3, 0),
+             numpy.timedelta64(60000000000, 'ns'), 24.0),
+            (datetime64(2014, 1, 1, 12, 4, 0),
+             numpy.timedelta64(60000000000, 'ns'), 4.5),
+        ], list(output))
+
+    def test_processor_derived_missing_boundary(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        tsb1.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 1, 0), 9),
+            (datetime64(2014, 1, 1, 12, 2, 0), 1),
+            (datetime64(2014, 1, 1, 12, 4, 0), 5),
+            (datetime64(2014, 1, 1, 12, 6, 0), 3)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc1))
+
+        tsb2.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 0, 0), 6),
+            (datetime64(2014, 1, 1, 12, 1, 0), 2),
+            (datetime64(2014, 1, 1, 12, 2, 0), 13),
+            (datetime64(2014, 1, 1, 12, 3, 0), 24),
+            (datetime64(2014, 1, 1, 12, 4, 0), 4)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc2))
+
+        output = processor.aggregated([
+            tsc1['return'], tsc2['return']], aggregation='mean',
+            from_timestamp=datetime64(2014, 1, 1, 12, 0, 0),
+            needed_percent_of_overlap=50.0)
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(60000000000, 'ns'), 6.0),
+            (datetime64(2014, 1, 1, 12, 1, 0),
+             numpy.timedelta64(60000000000, 'ns'), 5.5),
+            (datetime64(2014, 1, 1, 12, 2, 0),
+             numpy.timedelta64(60000000000, 'ns'), 7.0),
+            (datetime64(2014, 1, 1, 12, 3, 0),
+             numpy.timedelta64(60000000000, 'ns'), 24.0),
+            (datetime64(2014, 1, 1, 12, 4, 0),
+             numpy.timedelta64(60000000000, 'ns'), 4.5),
+        ], list(output))
+
+        output = processor.aggregated([
+            tsc1['return'], tsc2['return']], aggregation='mean',
+            to_timestamp=datetime64(2014, 1, 1, 12, 7, 0),
+            needed_percent_of_overlap=50.0)
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 12, 1, 0),
+             numpy.timedelta64(60000000000, 'ns'), 5.5),
+            (datetime64(2014, 1, 1, 12, 2, 0),
+             numpy.timedelta64(60000000000, 'ns'), 7.0),
+            (datetime64(2014, 1, 1, 12, 3, 0),
+             numpy.timedelta64(60000000000, 'ns'), 24.0),
+            (datetime64(2014, 1, 1, 12, 4, 0),
+             numpy.timedelta64(60000000000, 'ns'), 4.5),
+            (datetime64(2014, 1, 1, 12, 6, 0),
+             numpy.timedelta64(60000000000, 'ns'), 3.0),
+        ], list(output))
+
+    def test_aggregated_some_overlap_with_fill_zero(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        tsb1.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 3, 0), 9),
+            (datetime64(2014, 1, 1, 12, 4, 0), 1),
+            (datetime64(2014, 1, 1, 12, 5, 0), 2),
+            (datetime64(2014, 1, 1, 12, 6, 0), 7),
+            (datetime64(2014, 1, 1, 12, 7, 0), 5),
+            (datetime64(2014, 1, 1, 12, 8, 0), 3)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc1))
+
+        tsb2.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 0, 0), 6),
+            (datetime64(2014, 1, 1, 12, 1, 0), 2),
+            (datetime64(2014, 1, 1, 12, 2, 0), 13),
+            (datetime64(2014, 1, 1, 12, 3, 0), 24),
+            (datetime64(2014, 1, 1, 12, 4, 0), 4),
+            (datetime64(2014, 1, 1, 12, 5, 0), 16),
+            (datetime64(2014, 1, 1, 12, 6, 0), 12)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc2))
+
+        output = processor.aggregated([
+            tsc1['return'], tsc2['return']], aggregation='mean', fill=0)
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(60000000000, 'ns'), 3.0),
+            (datetime64(2014, 1, 1, 12, 1, 0),
+             numpy.timedelta64(60000000000, 'ns'), 1.0),
+            (datetime64(2014, 1, 1, 12, 2, 0),
+             numpy.timedelta64(60000000000, 'ns'), 6.5),
+            (datetime64(2014, 1, 1, 12, 3, 0),
+             numpy.timedelta64(60000000000, 'ns'), 16.5),
+            (datetime64(2014, 1, 1, 12, 4, 0),
+             numpy.timedelta64(60000000000, 'ns'), 2.5),
+            (datetime64(2014, 1, 1, 12, 5, 0),
+             numpy.timedelta64(60000000000, 'ns'), 9.0),
+            (datetime64(2014, 1, 1, 12, 6, 0),
+             numpy.timedelta64(60000000000, 'ns'), 9.5),
+            (datetime64(2014, 1, 1, 12, 7, 0),
+             numpy.timedelta64(60000000000, 'ns'), 2.5),
+            (datetime64(2014, 1, 1, 12, 8, 0),
+             numpy.timedelta64(60000000000, 'ns'), 1.5),
+        ], list(output))
+
+    def test_aggregated_some_overlap_with_fill_null(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        tsb1.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 3, 0), 9),
+            (datetime64(2014, 1, 1, 12, 4, 0), 1),
+            (datetime64(2014, 1, 1, 12, 5, 0), 2),
+            (datetime64(2014, 1, 1, 12, 6, 0), 7),
+            (datetime64(2014, 1, 1, 12, 7, 0), 5),
+            (datetime64(2014, 1, 1, 12, 8, 0), 3)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc1))
+
+        tsb2.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 0, 0), 6),
+            (datetime64(2014, 1, 1, 12, 1, 0), 2),
+            (datetime64(2014, 1, 1, 12, 2, 0), 13),
+            (datetime64(2014, 1, 1, 12, 3, 0), 24),
+            (datetime64(2014, 1, 1, 12, 4, 0), 4),
+            (datetime64(2014, 1, 1, 12, 5, 0), 16),
+            (datetime64(2014, 1, 1, 12, 6, 0), 12)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc2))
+
+        output = processor.aggregated([
+            tsc1['return'], tsc2['return']], aggregation='mean', fill='null')
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(60000000000, 'ns'), 6.0),
+            (datetime64(2014, 1, 1, 12, 1, 0),
+             numpy.timedelta64(60000000000, 'ns'), 2.0),
+            (datetime64(2014, 1, 1, 12, 2, 0),
+             numpy.timedelta64(60000000000, 'ns'), 13.0),
+            (datetime64(2014, 1, 1, 12, 3, 0),
+             numpy.timedelta64(60000000000, 'ns'), 16.5),
+            (datetime64(2014, 1, 1, 12, 4, 0),
+             numpy.timedelta64(60000000000, 'ns'), 2.5),
+            (datetime64(2014, 1, 1, 12, 5, 0),
+             numpy.timedelta64(60000000000, 'ns'), 9.0),
+            (datetime64(2014, 1, 1, 12, 6, 0),
+             numpy.timedelta64(60000000000, 'ns'), 9.5),
+            (datetime64(2014, 1, 1, 12, 7, 0),
+             numpy.timedelta64(60000000000, 'ns'), 5.0),
+            (datetime64(2014, 1, 1, 12, 8, 0),
+             numpy.timedelta64(60000000000, 'ns'), 3.0),
+        ], list(output))
+
+    def test_aggregate_no_points_with_fill_zero(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        tsb1.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 3, 0), 9),
+            (datetime64(2014, 1, 1, 12, 4, 0), 1),
+            (datetime64(2014, 1, 1, 12, 7, 0), 5),
+            (datetime64(2014, 1, 1, 12, 8, 0), 3)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc1))
+
+        tsb2.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 12, 0, 0), 6),
+            (datetime64(2014, 1, 1, 12, 1, 0), 2),
+            (datetime64(2014, 1, 1, 12, 2, 0), 13),
+            (datetime64(2014, 1, 1, 12, 3, 0), 24),
+            (datetime64(2014, 1, 1, 12, 4, 0), 4)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc2))
+
+        output = processor.aggregated([
+            tsc1['return'], tsc2['return']], aggregation='mean', fill=0)
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(60000000000, 'ns'), 3.0),
+            (datetime64(2014, 1, 1, 12, 1, 0),
+             numpy.timedelta64(60000000000, 'ns'), 1.0),
+            (datetime64(2014, 1, 1, 12, 2, 0),
+             numpy.timedelta64(60000000000, 'ns'), 6.5),
+            (datetime64(2014, 1, 1, 12, 3, 0),
+             numpy.timedelta64(60000000000, 'ns'), 16.5),
+            (datetime64(2014, 1, 1, 12, 4, 0),
+             numpy.timedelta64(60000000000, 'ns'), 2.5),
+            (datetime64(2014, 1, 1, 12, 7, 0),
+             numpy.timedelta64(60000000000, 'ns'), 2.5),
+            (datetime64(2014, 1, 1, 12, 8, 0),
+             numpy.timedelta64(60000000000, 'ns'), 1.5),
+        ], list(output))
+
+    def test_aggregated_nominal(self):
+        tsc1 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsc12 = {'sampling': numpy.timedelta64(300, 's'),
+                 'size': 6, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc12['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(60, 's'),
+                'size': 10, 'agg': 'mean'}
+        tsc22 = {'sampling': numpy.timedelta64(300, 's'),
+                 'size': 6, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc22['sampling'])
+
+        def ts1_update(ts):
+            grouped = ts.group_serie(tsc1['sampling'])
+            existing = tsc1.get('return')
+            tsc1['return'] = carbonara.AggregatedTimeSerie.from_grouped_serie(
+                grouped, tsc1['sampling'], tsc1['agg'],
+                max_size=tsc1['size'], truncate=True)
+            if existing:
+                existing.merge(tsc1['return'])
+            grouped = ts.group_serie(tsc12['sampling'])
+            existing = tsc12.get('return')
+            tsc12['return'] = carbonara.AggregatedTimeSerie.from_grouped_serie(
+                grouped, tsc12['sampling'], tsc12['agg'],
+                max_size=tsc12['size'], truncate=True)
+            if existing:
+                existing.merge(tsc12['return'])
+
+        def ts2_update(ts):
+            grouped = ts.group_serie(tsc2['sampling'])
+            existing = tsc2.get('return')
+            tsc2['return'] = carbonara.AggregatedTimeSerie.from_grouped_serie(
+                grouped, tsc2['sampling'], tsc2['agg'],
+                max_size=tsc2['size'], truncate=True)
+            if existing:
+                existing.merge(tsc2['return'])
+            grouped = ts.group_serie(tsc22['sampling'])
+            existing = tsc22.get('return')
+            tsc22['return'] = carbonara.AggregatedTimeSerie.from_grouped_serie(
+                grouped, tsc22['sampling'], tsc22['agg'],
+                max_size=tsc22['size'], truncate=True)
+            if existing:
+                existing.merge(tsc22['return'])
+        tsb1.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 11, 46, 4), 4),
+            (datetime64(2014, 1, 1, 11, 47, 34), 8),
+            (datetime64(2014, 1, 1, 11, 50, 54), 50),
+            (datetime64(2014, 1, 1, 11, 54, 45), 4),
+            (datetime64(2014, 1, 1, 11, 56, 49), 4),
+            (datetime64(2014, 1, 1, 11, 57, 22), 6),
+            (datetime64(2014, 1, 1, 11, 58, 22), 5),
+            (datetime64(2014, 1, 1, 12, 1, 4), 4),
+            (datetime64(2014, 1, 1, 12, 1, 9), 7),
+            (datetime64(2014, 1, 1, 12, 2, 1), 15),
+            (datetime64(2014, 1, 1, 12, 2, 12), 1),
+            (datetime64(2014, 1, 1, 12, 3, 0), 3),
+            (datetime64(2014, 1, 1, 12, 4, 9), 7),
+            (datetime64(2014, 1, 1, 12, 5, 1), 15),
+            (datetime64(2014, 1, 1, 12, 5, 12), 1),
+            (datetime64(2014, 1, 1, 12, 6, 0), 3)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=ts1_update)
+
+        tsb2.set_values(numpy.array([
+            (datetime64(2014, 1, 1, 11, 46, 4), 6),
+            (datetime64(2014, 1, 1, 11, 47, 34), 5),
+            (datetime64(2014, 1, 1, 11, 50, 54), 51),
+            (datetime64(2014, 1, 1, 11, 54, 45), 5),
+            (datetime64(2014, 1, 1, 11, 56, 49), 5),
+            (datetime64(2014, 1, 1, 11, 57, 22), 7),
+            (datetime64(2014, 1, 1, 11, 58, 22), 5),
+            (datetime64(2014, 1, 1, 12, 1, 4), 5),
+            (datetime64(2014, 1, 1, 12, 1, 9), 8),
+            (datetime64(2014, 1, 1, 12, 2, 1), 10),
+            (datetime64(2014, 1, 1, 12, 2, 12), 2),
+            (datetime64(2014, 1, 1, 12, 3, 0), 6),
+            (datetime64(2014, 1, 1, 12, 4, 9), 4),
+            (datetime64(2014, 1, 1, 12, 5, 1), 10),
+            (datetime64(2014, 1, 1, 12, 5, 12), 1),
+            (datetime64(2014, 1, 1, 12, 6, 0), 1)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=ts2_update)
+        output = processor.aggregated(
+            [tsc1['return'], tsc12['return'], tsc2['return'], tsc22['return']],
+            'mean')
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 11, 45),
+             numpy.timedelta64(300, 's'), 5.75),
+            (datetime64(2014, 1, 1, 11, 50),
+             numpy.timedelta64(300, 's'), 27.5),
+            (datetime64(2014, 1, 1, 11, 55),
+             numpy.timedelta64(300, 's'), 5.3333333333333339),
+            (datetime64(2014, 1, 1, 12, 0),
+             numpy.timedelta64(300, 's'), 6.0),
+            (datetime64(2014, 1, 1, 12, 5),
+             numpy.timedelta64(300, 's'), 5.1666666666666661),
+            (datetime64(2014, 1, 1, 11, 54),
+             numpy.timedelta64(60, 's'), 4.5),
+            (datetime64(2014, 1, 1, 11, 56),
+             numpy.timedelta64(60, 's'), 4.5),
+            (datetime64(2014, 1, 1, 11, 57),
+             numpy.timedelta64(60, 's'), 6.5),
+            (datetime64(2014, 1, 1, 11, 58),
+             numpy.timedelta64(60, 's'), 5.0),
+            (datetime64(2014, 1, 1, 12, 1),
+             numpy.timedelta64(60, 's'), 6.0),
+            (datetime64(2014, 1, 1, 12, 2),
+             numpy.timedelta64(60, 's'), 7.0),
+            (datetime64(2014, 1, 1, 12, 3),
+             numpy.timedelta64(60, 's'), 4.5),
+            (datetime64(2014, 1, 1, 12, 4),
+             numpy.timedelta64(60, 's'), 5.5),
+            (datetime64(2014, 1, 1, 12, 5),
+             numpy.timedelta64(60, 's'), 6.75),
+            (datetime64(2014, 1, 1, 12, 6),
+             numpy.timedelta64(60, 's'), 2.0),
+        ], list(output))
+
+    def test_aggregated_partial_overlap(self):
+        tsc1 = {'sampling': numpy.timedelta64(1, 's'),
+                'size': 86400, 'agg': 'mean'}
+        tsb1 = carbonara.BoundTimeSerie(block_size=tsc1['sampling'])
+        tsc2 = {'sampling': numpy.timedelta64(1, 's'),
+                'size': 60, 'agg': 'mean'}
+        tsb2 = carbonara.BoundTimeSerie(block_size=tsc2['sampling'])
+
+        tsb1.set_values(numpy.array([
+            (datetime64(2015, 12, 3, 13, 19, 15), 1),
+            (datetime64(2015, 12, 3, 13, 20, 15), 1),
+            (datetime64(2015, 12, 3, 13, 21, 15), 1),
+            (datetime64(2015, 12, 3, 13, 22, 15), 1)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc1))
+
+        tsb2.set_values(numpy.array([
+            (datetime64(2015, 12, 3, 13, 21, 15), 10),
+            (datetime64(2015, 12, 3, 13, 22, 15), 10),
+            (datetime64(2015, 12, 3, 13, 23, 15), 10),
+            (datetime64(2015, 12, 3, 13, 24, 15), 10)],
+            dtype=carbonara.TIMESERIES_ARRAY_DTYPE),
+            before_truncate_callback=functools.partial(
+                self._resample_and_merge, agg_dict=tsc2))
+
+        output = processor.aggregated(
+            [tsc1['return'], tsc2['return']], aggregation="sum")
+
+        self.assertEqual([
+            (datetime64(
+                2015, 12, 3, 13, 21, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+            (datetime64(
+                2015, 12, 3, 13, 22, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+        ], list(output))
+
+        dtfrom = datetime64(2015, 12, 3, 13, 17, 0)
+        dtto = datetime64(2015, 12, 3, 13, 25, 0)
+
+        output = processor.aggregated(
+            [tsc1['return'], tsc2['return']],
+            from_timestamp=dtfrom, to_timestamp=dtto,
+            aggregation="sum", needed_percent_of_overlap=0)
+        self.assertEqual([
+            (datetime64(
+                2015, 12, 3, 13, 19, 15
+            ), numpy.timedelta64(1, 's'), 1.0),
+            (datetime64(
+                2015, 12, 3, 13, 20, 15
+            ), numpy.timedelta64(1, 's'), 1.0),
+            (datetime64(
+                2015, 12, 3, 13, 21, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+            (datetime64(
+                2015, 12, 3, 13, 22, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+            (datetime64(
+                2015, 12, 3, 13, 23, 15
+            ), numpy.timedelta64(1, 's'), 10.0),
+            (datetime64(
+                2015, 12, 3, 13, 24, 15
+            ), numpy.timedelta64(1, 's'), 10.0),
+        ], list(output))
+
+        # Check boundaries are set when overlap=0
+        output = processor.aggregated(
+            [tsc1['return'], tsc2['return']],
+            aggregation="sum", needed_percent_of_overlap=0)
+        self.assertEqual([
+            (datetime64(
+                2015, 12, 3, 13, 21, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+            (datetime64(
+                2015, 12, 3, 13, 22, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+        ], list(output))
+
+        # By default we require 100% of point that overlap
+        # so that fail if from or to is set
+        self.assertRaises(processor.UnAggregableTimeseries,
+                          processor.aggregated,
+                          [tsc1['return'], tsc2['return']],
+                          to_timestamp=dtto, aggregation='mean')
+        self.assertRaises(processor.UnAggregableTimeseries,
+                          processor.aggregated,
+                          [tsc1['return'], tsc2['return']],
+                          from_timestamp=dtfrom, aggregation='mean')
+        # Retry with 50% and it works
+        output = processor.aggregated(
+            [tsc1['return'], tsc2['return']], from_timestamp=dtfrom,
+            aggregation="sum",
+            needed_percent_of_overlap=50.0)
+        self.assertEqual([
+            (datetime64(
+                2015, 12, 3, 13, 19, 15
+            ), numpy.timedelta64(1, 's'), 1.0),
+            (datetime64(
+                2015, 12, 3, 13, 20, 15
+            ), numpy.timedelta64(1, 's'), 1.0),
+            (datetime64(
+                2015, 12, 3, 13, 21, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+            (datetime64(
+                2015, 12, 3, 13, 22, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+        ], list(output))
+
+        output = processor.aggregated(
+            [tsc1['return'], tsc2['return']], to_timestamp=dtto,
+            aggregation="sum",
+            needed_percent_of_overlap=50.0)
+        self.assertEqual([
+            (datetime64(
+                2015, 12, 3, 13, 21, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+            (datetime64(
+                2015, 12, 3, 13, 22, 15
+            ), numpy.timedelta64(1, 's'), 11.0),
+            (datetime64(
+                2015, 12, 3, 13, 23, 15
+            ), numpy.timedelta64(1, 's'), 10.0),
+            (datetime64(
+                2015, 12, 3, 13, 24, 15
+            ), numpy.timedelta64(1, 's'), 10.0),
+        ], list(output))
+
+
+class CrossMetricAggregated(base.TestCase):
     def setUp(self):
-        super(TestAggregates, self).setUp()
-        mgr = extension.ExtensionManager('gnocchi.aggregates',
-                                         invoke_on_load=True)
-        self.custom_agg = dict((x.name, x.obj) for x in mgr)
+        super(CrossMetricAggregated, self).setUp()
+        # A lot of tests wants a metric, create one
+        self.metric, __ = self._create_metric()
 
-    def test_extension_dict(self):
-        self.assertIsInstance(self.custom_agg['moving-average'],
-                              moving_stats.MovingAverage)
+    def test_get_measures_empty_metrics_no_overlap(self):
+        self.assertRaises(
+            processor.MetricUnaggregatable,
+            processor.get_measures, self.storage,
+            [indexer.Metric(uuid.uuid4(), self.archive_policies['low']),
+             indexer.Metric(uuid.uuid4(), self.archive_policies['low'])])
 
-    def _test_create_metric_and_data(self, data, spacing):
-        metric = indexer.Metric(
-            uuid.uuid4(), self.archive_policies['medium'])
-        start_time = utils.datetime_utc(2014, 1, 1, 12)
-        incr = datetime.timedelta(seconds=spacing)
-        measures = [incoming.Measure(
-            utils.dt_in_unix_ns(start_time + incr * n), val)
-            for n, val in enumerate(data)]
-        self.index.create_metric(metric.id, str(uuid.uuid4()), 'medium')
-        self.incoming.add_measures(metric, measures)
-        metrics = tests_utils.list_all_incoming_metrics(self.incoming)
-        self.storage.process_new_measures(
-            self.index, self.incoming, metrics, sync=True)
+    def test_get_measures_unknown_aggregation(self):
+        metric2 = indexer.Metric(uuid.uuid4(),
+                                 self.archive_policies['low'])
+        self.incoming.add_measures(self.metric, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 31), 42),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 9, 31), 4),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 12, 45), 44),
+        ])
+        self.incoming.add_measures(metric2, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 31), 42),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 9, 31), 4),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 12, 45), 44),
+        ])
+        self.assertRaises(storage.AggregationDoesNotExist,
+                          processor.get_measures,
+                          self.storage,
+                          [self.metric, metric2],
+                          aggregation='last')
 
-        return metric
+    def test_get_measures_unknown_granularity(self):
+        metric2 = indexer.Metric(uuid.uuid4(),
+                                 self.archive_policies['low'])
+        self.incoming.add_measures(self.metric, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 31), 42),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 9, 31), 4),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 12, 45), 44),
+        ])
+        self.incoming.add_measures(metric2, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 31), 42),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 9, 31), 4),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 12, 45), 44),
+        ])
+        self.assertRaises(storage.GranularityDoesNotExist,
+                          processor.get_measures,
+                          self.storage,
+                          [self.metric, metric2],
+                          granularity=numpy.timedelta64(12345456, 'ms'))
 
-    def test_retrieve_data(self):
-        metric = self._test_create_metric_and_data([69, 42, 6, 44, 7],
-                                                   spacing=20)
-        for agg_method in self.custom_agg:
-            agg_obj = self.custom_agg[agg_method]
-            window = numpy.timedelta64(90, 's')
-            self.assertRaises(aggregates.CustomAggFailure,
-                              agg_obj.retrieve_data,
-                              self.storage, metric,
-                              start=None, stop=None,
-                              window=window)
+    def test_add_and_get_measures_different_archives(self):
+        metric2 = indexer.Metric(uuid.uuid4(),
+                                 self.archive_policies['no_granularity_match'])
+        self.incoming.add_measures(self.metric, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 31), 42),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 9, 31), 4),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 12, 45), 44),
+        ])
+        self.incoming.add_measures(metric2, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 31), 42),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 9, 31), 4),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 12, 45), 44),
+        ])
 
-            window = numpy.timedelta64(120, 's')
-            grain, result = agg_obj.retrieve_data(self.storage, metric,
-                                                  start=None, stop=None,
-                                                  window=window)
-            self.assertEqual(numpy.timedelta64(1, 'm'), grain)
-            self.assertEqual(39.0, result[datetime.datetime(2014, 1, 1, 12)])
-            self.assertEqual(25.5,
-                             result[datetime.datetime(2014, 1, 1, 12, 1)])
-        self.storage.delete_metric(self.incoming, metric)
+        self.assertRaises(processor.MetricUnaggregatable,
+                          processor.get_measures,
+                          self.storage,
+                          [self.metric, metric2])
 
-    def test_compute_moving_average(self):
-        metric = self._test_create_metric_and_data([69, 42, 6, 44, 7],
-                                                   spacing=20)
-        agg_obj = self.custom_agg['moving-average']
-        window = '120s'
+    def test_add_and_get_measures(self):
+        metric2, __ = self._create_metric()
+        self.incoming.add_measures(self.metric, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 31), 42),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 9, 31), 4),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 12, 45), 44),
+        ])
+        self.incoming.add_measures(metric2, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 5), 9),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 41), 2),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 10, 31), 4),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 13, 10), 4),
+        ])
+        self.trigger_processing([str(self.metric.id), str(metric2.id)])
 
-        center = 'False'
-        result = agg_obj.compute(self.storage, metric,
-                                 start=None, stop=None,
-                                 window=window, center=center)
-        self.assertEqual([(numpy.datetime64("2014-01-01 12:00"),
-                           numpy.timedelta64(120, 's'),
-                           32.25)],
-                         result)
+        values = processor.get_measures(
+            self.storage, [self.metric, metric2])
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 0, 0, 0),
+             numpy.timedelta64(1, 'D'), 22.25),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(1, 'h'), 22.25),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(5, 'm'), 39.0),
+            (datetime64(2014, 1, 1, 12, 5, 0),
+             numpy.timedelta64(5, 'm'), 12.5),
+            (datetime64(2014, 1, 1, 12, 10, 0),
+             numpy.timedelta64(5, 'm'), 24.0)
+        ], values)
 
-        center = 'True'
-        result = agg_obj.compute(self.storage, metric,
-                                 start=None, stop=None,
-                                 window=window, center=center)
+        values = processor.get_measures(
+            self.storage, [self.metric, metric2], reaggregation='max')
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 0, 0, 0),
+             numpy.timedelta64(1, 'D'), 39.75),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(1, 'h'), 39.75),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(5, 'm'), 69),
+            (datetime64(2014, 1, 1, 12, 5, 0),
+             numpy.timedelta64(5, 'm'), 23),
+            (datetime64(2014, 1, 1, 12, 10, 0),
+             numpy.timedelta64(5, 'm'), 44)
+        ], values)
 
-        self.assertEqual([(numpy.datetime64("2014-01-01 12:01"),
-                           numpy.timedelta64(120, 's'),
-                           28.875)],
-                         result)
-        # (FIXME) atmalagon: doing a centered average when
-        # there are only two points in the retrieved data seems weird.
-        # better to raise an error or return nan in this case?
+        values = processor.get_measures(
+            self.storage, [self.metric, metric2],
+            from_timestamp=datetime64(2014, 1, 1, 12, 10, 0))
+        self.assertEqual([
+            (datetime64(2014, 1, 1),
+             numpy.timedelta64(1, 'D'), 22.25),
+            (datetime64(2014, 1, 1, 12),
+             numpy.timedelta64(1, 'h'), 22.25),
+            (datetime64(2014, 1, 1, 12, 10, 0),
+             numpy.timedelta64(5, 'm'), 24.0),
+        ], values)
 
-        self.storage.delete_metric(self.incoming, metric)
+        values = processor.get_measures(
+            self.storage, [self.metric, metric2],
+            to_timestamp=datetime64(2014, 1, 1, 12, 5, 0))
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 0, 0, 0),
+             numpy.timedelta64(1, 'D'), 22.25),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(1, 'h'), 22.25),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(5, 'm'), 39.0),
+        ], values)
+
+        values = processor.get_measures(
+            self.storage, [self.metric, metric2],
+            from_timestamp=datetime64(2014, 1, 1, 12, 10, 10),
+            to_timestamp=datetime64(2014, 1, 1, 12, 10, 10))
+        self.assertEqual([
+            (datetime64(2014, 1, 1),
+             numpy.timedelta64(1, 'D'), 22.25),
+            (datetime64(2014, 1, 1, 12),
+             numpy.timedelta64(1, 'h'), 22.25),
+            (datetime64(2014, 1, 1, 12, 10),
+             numpy.timedelta64(5, 'm'), 24.0),
+        ], values)
+
+        values = processor.get_measures(
+            self.storage, [self.metric, metric2],
+            from_timestamp=datetime64(2014, 1, 1, 12, 0, 0),
+            to_timestamp=datetime64(2014, 1, 1, 12, 0, 1))
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1),
+             numpy.timedelta64(1, 'D'), 22.25),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(1, 'h'), 22.25),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(5, 'm'), 39.0),
+        ], values)
+
+        values = processor.get_measures(
+            self.storage, [self.metric, metric2],
+            from_timestamp=datetime64(2014, 1, 1, 12, 0, 0),
+            to_timestamp=datetime64(2014, 1, 1, 12, 0, 1),
+            granularity=numpy.timedelta64(5, 'm'))
+
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(5, 'm'), 39.0),
+        ], values)
+
+    def test_add_and_get_measures_with_holes(self):
+        metric2, __ = self._create_metric()
+        self.incoming.add_measures(self.metric, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 1), 69),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 31), 42),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 5, 31), 8),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 9, 31), 4),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 12, 45), 42),
+        ])
+        self.incoming.add_measures(metric2, [
+            incoming.Measure(datetime64(2014, 1, 1, 12, 0, 5), 9),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 7, 31), 2),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 9, 31), 6),
+            incoming.Measure(datetime64(2014, 1, 1, 12, 13, 10), 2),
+        ])
+        self.trigger_processing([str(self.metric.id), str(metric2.id)])
+
+        values = processor.get_measures(
+            self.storage, [self.metric, metric2])
+        self.assertEqual([
+            (datetime64(2014, 1, 1, 0, 0, 0),
+             numpy.timedelta64(1, 'D'), 18.875),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(1, 'h'), 18.875),
+            (datetime64(2014, 1, 1, 12, 0, 0),
+             numpy.timedelta64(5, 'm'), 39.0),
+            (datetime64(2014, 1, 1, 12, 5, 0),
+             numpy.timedelta64(5, 'm'), 11.0),
+            (datetime64(2014, 1, 1, 12, 10, 0),
+             numpy.timedelta64(5, 'm'), 22.0)
+        ], values)
