+++ setup.cfg
@@ -112,7 +112,7 @@ gnocchi.indexer =
     postgresql+psycopg2 = gnocchi.indexer.sqlalchemy:SQLAlchemyIndexer
 
 gnocchi.aggregates =
-    moving-average = gnocchi.aggregates.moving_stats:MovingAverage
+    moving-average = gnocchi.deprecated_aggregates.moving_stats:MovingAverage
 
 gnocchi.rest.auth_helper =
     keystone = gnocchi.rest.auth_helper:KeystoneAuthHelper
